Replace dedicated_disks by storage_dir.
Currently we don't manage disks. So instead we offer the option to use
the same lvm than the one created by the Elasticsearch-Kibana plugin. To
be able to use this we need to add a dependency between the two plugins and
run the installation of InfluxDB after the creation of the LVM.

Change-Id: I305c01ede084fc170335cef24633a9e9dba0eb1d