Expose storage_dir from puppet manifest.
Currently we don't manage disks. Instead, we offer the option to use
the same LVM than the one created by the Elasticsearch-Kibana plugin.
As we are not sure about how two plugins are evaluated by Fuel, we
keep this option hidden until we can manage the deployment of several
plugins.

Change-Id: I305c01ede084fc170335cef24633a9e9dba0eb1d