Remove dependency on cluster-haproxy.pp from osnailyfacter
This patch adds a new manifest to replace the one that is used from
osnailyfacter. We cannot use the manifest provided by fuel-library
because it has been rewritten and it is not compatible with our
deployment: it assumes a fixed VIP naming schema that is specific
to the controllers.

Partial-Bug: #1554521
Change-Id: Ibcac60dc89787500b931296e94739e692b544cdd