Remove dependency on cluster-haproxy.pp from osnailyfacter
This patch adds a new manifest to replace the one that is used from
osnailyfacter. We cannot use the manifest provided by fuel-library
because it has been rewritten and it is not compatible with our
deployment because it it requires specific VIP that are tied to the
management and public VIP addresses of the controllers.

Partial-Bug: #1554521
Change-Id: Ibcac60dc89787500b931296e94739e692b544cdd