Update to task based deployment
MOS 9 comes with a new orchestration engine (so-called v2 or task-based
deployement) which reduces greatly the time necessary to deploy new
environments.

This change updates the deployment tasks to be fully compatible with
the v2 engine since when a deployment mixes v1 and v2 tasks, the engine
falls back to the v1 (slower) mode.

Note that we need to keep the backward compatibility with MOS 8 so the
deployment tasks still define the v1 attributes like 'requires' and
'required_for'.

Change-Id: I5df3e649b2f31f2e1a5ccd59f346d870651315d5