Fix tasks dependency when using remote backends
Previously the LMA collector had an explicit dependency on the last
InfluxDB task but this doesn't work when the InfluxDB/Grafana
plugin isn't deployed in the same environment as the collector plugin.

This change makes sure that the LMA collector plugin's tasks will
only be executed once the InfluxDB cluster is fully functional by
pinning both plugins to a well-known task (update_hosts in that case).

Partial-Bug: #1573087
Closes-Bug: #1578194
Change-Id: I830f024e0d1629ae10a8a504c4acbddf23d5c47e