Fix dependency for grafana-configuration task
This change removes an unneeded dependency and it fixes a wrong one.
We don't need to wait that Grafana is configured before pushing data
to the backend, so we removed the dependency on upload_nodes_info.
And as the task grafana-configuration is only executed on the primary
role it depends on primary-grafana-installation.

Closes-Bug: #1585112
Change-Id: Id79eb6680ca725b6f60b82d0e9d43acc61b35fe1