Fix a cycle in graph of tasks
This patch removes a dependency that causes a cycle in the graph of
tasks in MOS 8. The cycle was <custom>-netconfig -> hosts -> cluster
-> virtual_ips -> <custom>-netconfig.

We cannot remove hosts.pp as we tried previously because doing this also
remove a dependency between globals and hosts. So we create the
following dependency:

tools -> <custom>-netconfig -> hosts

in place of the former:

tools -> netconfig -> hosts

Change-Id: I67abb9f45a0156017236defb04196c701f75926b