Add the option to store the InfluxDB WAL in memory
This change adds a new option to the Fuel UI to store the InfluxDB WAL
files in memory. This greatly improves write performances but it may
lead to data loss in case of server crash.

The WAL partition will use 10% of the total RAM (but no more than 4GB).

DocImpact

Change-Id: I6153cd24322ef3eab53505bf4cb1f72d9594ec04